#54 [refactor] 카카오 로그인 구현 부분 ActivityContext로 캐스팅해서 주입받도록 힐트 적용하여 리팩토링
