+++ app/src/main/java/com/spark/android/data/repository/SignInRepository.kt
@@ -1,5 +1,7 @@
 package com.spark.android.data.repository
 
+import com.kakao.sdk.auth.model.OAuthToken
+
 interface SignInRepository {
-    fun startKakaoLogin():Int
+    fun startKakaoLogin(kakaoLoginCallback: (OAuthToken?, Throwable?) -> Unit)
 }
