+++ app/src/main/java/com/spark/android/data/repository/SignInRepositoryImpl.kt
@@ -1,21 +1,32 @@
 package com.spark.android.data.repository
 
 import android.content.Context
-import android.util.Log
 import com.kakao.sdk.auth.model.OAuthToken
 import com.kakao.sdk.user.UserApiClient
+import dagger.hilt.android.qualifiers.ActivityContext
 import javax.inject.Inject
 
 class SignInRepositoryImpl @Inject constructor(
-    private val context: Context
+    @ActivityContext private val context: Context
 ) : SignInRepository {
-    override fun startKakaoLogin(): Int =
-        if (UserApiClient.instance.isKakaoTalkLoginAvailable(context)) {
+    override fun startKakaoLogin(kakaoLoginCallback: (OAuthToken?, Throwable?) -> Unit) {
+        val kakaoLoginState = if (UserApiClient.instance.isKakaoTalkLoginAvailable(context)) {
             KAKAO_TALK_LOGIN
         } else {
             KAKAO_ACCOUNT_LOGIN
         }
 
+        when (kakaoLoginState) {
+            KAKAO_TALK_LOGIN -> UserApiClient.instance.loginWithKakaoTalk(
+                context,
+                callback = kakaoLoginCallback
+            )
+            KAKAO_ACCOUNT_LOGIN -> UserApiClient.instance.loginWithKakaoAccount(
+                context,
+                callback = kakaoLoginCallback
+            )
+        }
+    }
 
     companion object {
         const val KAKAO_TALK_LOGIN = 0
