+++ app/src/main/java/com/spark/android/ui/signin/SignInFragment.kt
@@ -1,18 +1,16 @@
 package com.spark.android.ui.signin
 
 import android.os.Bundle
-import android.util.Log
 import android.view.View
 import androidx.fragment.app.viewModels
-import com.kakao.sdk.user.UserApiClient
 import com.spark.android.R
-import com.spark.android.data.repository.SignInRepositoryImpl.Companion.KAKAO_ACCOUNT_LOGIN
-import com.spark.android.data.repository.SignInRepositoryImpl.Companion.KAKAO_TALK_LOGIN
 import com.spark.android.databinding.FragmentSignInBinding
 import com.spark.android.ui.base.BaseFragment
+import com.spark.android.util.EventObserver
 import com.spark.android.util.initStatusBarColor
 import com.spark.android.util.initStatusBarTextColorToWhite
 import com.spark.android.util.navigate
+import com.spark.android.util.showToast
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -23,8 +21,7 @@ class SignInFragment : BaseFragment<FragmentSignInBinding>(R.layout.fragment_sig
         super.onViewCreated(view, savedInstanceState)
         binding.signInViewModel = signInViewModel
         initStatusBarStyle()
-        initKakaoLoginStateObserver()
-        initKakaoUserIdObserver()
+        initIsSuccessKakaoLoginObserver()
     }
 
     private fun initStatusBarStyle() {
@@ -32,40 +29,13 @@ class SignInFragment : BaseFragment<FragmentSignInBinding>(R.layout.fragment_sig
         requireActivity().initStatusBarTextColorToWhite()
     }
 
-    private fun initKakaoLoginStateObserver() {
-        signInViewModel.kakaoLoginState.observe(viewLifecycleOwner) { kakaoLoginState ->
-            when (kakaoLoginState) {
-                KAKAO_TALK_LOGIN -> UserApiClient.instance.loginWithKakaoTalk(
-                    requireContext(),
-                    callback = signInViewModel.kakaoLoginCallback
-                )
-                KAKAO_ACCOUNT_LOGIN -> UserApiClient.instance.loginWithKakaoAccount(
-                    requireContext(),
-                    callback = signInViewModel.kakaoLoginCallback
-                )
+    private fun initIsSuccessKakaoLoginObserver() {
+        signInViewModel.isSuccessKakaoLogin.observe(viewLifecycleOwner, EventObserver { isSuccess ->
+            if (isSuccess) {
+                navigate(R.id.action_signInFragment_to_profileFragment)
+            } else {
+                requireContext().showToast(getString(R.string.sign_in_kakao_login_failure_msg))
             }
-            initKakaoUserId()
-        }
-    }
-
-    private fun initKakaoUserIdObserver() {
-        signInViewModel.kakaoUserId.observe(viewLifecycleOwner) {
-            navigate(R.id.action_signInFragment_to_profileFragment)
-        }
-    }
-
-    private fun initKakaoUserId() {
-        UserApiClient.instance.accessTokenInfo { tokenInfo, error ->
-            if (error != null) {
-                Log.e("kakao", "토큰 정보 보기 실패", error)
-            } else if (tokenInfo != null) {
-                signInViewModel.initKakaoUserId(requireNotNull(tokenInfo.id))
-                Log.d(
-                    "kakao", "토큰 정보 보기 성공" +
-                            "\n회원번호: ${tokenInfo.id}" +
-                            "\n만료시간: ${tokenInfo.expiresIn} 초"
-                )
-            }
-        }
+        })
     }
 }
