+++ app/src/main/java/com/spark/android/ui/signin/SignInViewModel.kt
@@ -6,6 +6,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.kakao.sdk.auth.model.OAuthToken
 import com.spark.android.data.repository.SignInRepository
+import com.spark.android.util.Event
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
@@ -13,26 +14,24 @@ import javax.inject.Inject
 class SignInViewModel @Inject constructor(
     private val signInRepository: SignInRepository
 ) : ViewModel() {
-    private val _kakaoLoginState = MutableLiveData<Int>()
-    val kakaoLoginState: LiveData<Int> = _kakaoLoginState
-
-    private val _kakaoUserId = MutableLiveData<Long>()
-    val kakaoUserId: LiveData<Long> = _kakaoUserId
-
+    private val _isSuccessKakaoLogin = MutableLiveData<Event<Boolean>>()
+    val isSuccessKakaoLogin: LiveData<Event<Boolean>> = _isSuccessKakaoLogin
 
     val kakaoLoginCallback: (OAuthToken?, Throwable?) -> Unit = { token, error ->
         if (error != null) {
+            initIsSuccessLogin(false)
             Log.e("kakao", "로그인 실패", error)
         } else if (token != null) {
+            initIsSuccessLogin(true)
             Log.d("kakao", "로그인 성공 ${token.accessToken}")
         }
     }
 
     fun initKakaoLoginState() {
-        _kakaoLoginState.value = signInRepository.startKakaoLogin()
+        signInRepository.startKakaoLogin(kakaoLoginCallback)
     }
 
-    fun initKakaoUserId(userId: Long) {
-        _kakaoUserId.value = userId
+    private fun initIsSuccessLogin(isSuccess: Boolean) {
+        _isSuccessKakaoLogin.value = Event(isSuccess)
     }
 }
